backend/requirements.txt
fastapi~=0.109.0
fastapi-cli
uvicorn[standard]~=0.22.0
pydantic~=2.6.0
passlib[argon2]~=1.7.4
python-jose==3.3.0
python-multipart
websockets~=12.0
sqlalchemy~=2.0.36
paho-mqtt~=2.1.0
PyYAML
numpy
jinja2